# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Brand:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      order_number:
        type: integer
      parent_id:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.City:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  utils.DataResponse:
    properties:
      error_code:
        type: integer
      message: {}
    type: object
host: localhost:5005
info:
  contact: {}
  description: Api INSIGHT project
  title: Insight API SERVICE
  version: 1.0.0
paths:
...
  /settings/brands:
    delete:
      consumes:
      - application/json
      description: Удаление бренда
      operationId: deleteBrand
      parameters:
      - description: Введите данные
        in: query
        name: brand_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteBrand
      tags:
      - Settings
    get:
      consumes:
      - application/json
      description: Просмотр всех брендов
      operationId: getAllBrands
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllBrands
      tags:
      - Settings
    post:
      consumes:
      - application/json
      description: Добавление бренда
      operationId: addBrand
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addBrand
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Изменение бренда
      operationId: editBrand
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editBrand
      tags:
      - Settings
  /settings/categories:
    delete:
      consumes:
      - application/json
      description: Удаление категории
      operationId: deleteCategory
      parameters:
      - description: Введите данные
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteCategory
      tags:
      - Settings
    get:
      consumes:
      - application/json
      description: Просмотр всех категорий
      operationId: getAllCategories
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllCategories
      tags:
      - Settings
    post:
      consumes:
      - application/json
      description: Добавление категории
      operationId: addNewCategory
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewCategory
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Редактирование категории
      operationId: editCategory
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editCategory
      tags:
      - Settings
  /settings/cities:
    delete:
      consumes:
      - application/json
      description: Удаление города
      operationId: deleteCity
      parameters:
      - description: Введите данные
        in: query
        name: city_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteCity
      tags:
      - Settings
    get:
      consumes:
      - application/json
      description: Просмотр всех городов
      operationId: getAllCities
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllCities
      tags:
      - Settings
    post:
      consumes:
      - application/json
      description: Добавление города
      operationId: addNewCity
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewCity
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Редактирование города
      operationId: editCity
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editCity
      tags:
      - Settings
  /settings/permissions:
    get:
      consumes:
      - application/json
      description: Просмотр всех доступов
      operationId: getAllPermissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllPermissions
      tags:
      - Settings
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"